feat(TI-CORE:1): consolidate runtime and domain modules
