+++ __AGENT__/AGENT_DECISIONS.md
@@ -1,1 +1,2 @@
 2025-09-24: Assumed 'continue' meant advancing modular refactor by isolating camera/lens controls.
+2025-09-26: Assumed TI plan approval covered full runtime/domain consolidation into core/runtime + src/domain.
