+++ __AGENT__/AGENT_METRICS.md
@@ -1,2 +1,3 @@
 2025-09-24: Task - Modular camera controller extraction; Result - completed with build passing; Next - extend modularization to materials/post presets.
 2025-09-24: Task - Config + stage refactor kickoff; Result - schema established, stage pipeline migrated, diagnostics updated; Next - migrate remaining modules to new config & implement plugin host.
+2025-09-26: Task - TI consolidation runtime/domain restructure; Result - runtime unified with domain factories + examples, lint/build green; Next - monitor domain-specific follow-ups.
