+++ tsconfig.json
@@ -19,7 +19,7 @@
     "baseUrl": ".",
     "paths": {
       "@core/*": ["src/core/*"],
-      "@modules/*": ["src/modules/*"],
+      "@modules/*": ["src/domain/*"],
       "@presets/*": ["src/presets/*"]
     }
   },
